Use custom resolver for createProject mutation
