+++ tests/Feature/GraphQL/ProjectTypeTest.php
@@ -5,7 +5,6 @@ namespace Tests\Feature\GraphQL;
 use App\Models\Project;
 use App\Models\User;
 use Illuminate\Foundation\Testing\DatabaseTruncation;
-use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Str;
 use PHPUnit\Framework\Attributes\DataProvider;
 use Tests\TestCase;
@@ -335,156 +334,6 @@ class ProjectTypeTest extends TestCase
                 ],
             ],
         ]);
-    }
-
-    public function testCreateProjectNoUser(): void
-    {
-        $name = 'test-project' . Str::uuid();
-        $this->graphQL('
-            mutation CreateProject($input: CreateProjectInput!) {
-                createProject(input: $input) {
-                    id
-                    name
-                }
-            }
-        ', [
-            'input' => [
-                'name' => $name,
-                'description' => 'test',
-                'homeUrl' => 'https://cdash.org',
-                'visibility' => 'PUBLIC',
-                'authenticateSubmissions' => false,
-            ],
-        ])->assertGraphQLErrorMessage('This action is unauthorized.');
-
-        // A final check to ensure this project wasn't created anyway
-        $this->assertDatabaseMissing(Project::class, [
-            'name' => $name,
-        ]);
-    }
-
-    public function testCreateProjectUnauthorizedUser(): void
-    {
-        $name = 'test-project' . Str::uuid();
-        $this->actingAs($this->users['normal'])->graphQL('
-            mutation CreateProject($input: CreateProjectInput!) {
-                createProject(input: $input) {
-                    id
-                    name
-                }
-            }
-        ', [
-            'input' => [
-                'name' => $name,
-                'description' => 'test',
-                'homeUrl' => 'https://cdash.org',
-                'visibility' => 'PUBLIC',
-                'authenticateSubmissions' => false,
-            ],
-        ])->assertGraphQLErrorMessage('This action is unauthorized.');
-
-        // A final check to ensure this project wasn't created anyway
-        $this->assertDatabaseMissing(Project::class, [
-            'name' => $name,
-        ]);
-    }
-
-    public function testCreateProjectUserCreateProjectNoUser(): void
-    {
-        Config::set('cdash.user_create_projects', true);
-
-        $name = 'test-project' . Str::uuid();
-        $this->graphQL('
-            mutation CreateProject($input: CreateProjectInput!) {
-                createProject(input: $input) {
-                    id
-                    name
-                }
-            }
-        ', [
-            'input' => [
-                'name' => $name,
-                'description' => 'test',
-                'homeUrl' => 'https://cdash.org',
-                'visibility' => 'PUBLIC',
-                'authenticateSubmissions' => false,
-            ],
-        ])->assertGraphQLErrorMessage('This action is unauthorized.');
-
-        // A final check to ensure this project wasn't created anyway
-        $this->assertDatabaseMissing(Project::class, [
-            'name' => $name,
-        ]);
-    }
-
-    public function testCreateProjectUserCreateProject(): void
-    {
-        Config::set('cdash.user_create_projects', true);
-
-        $name = 'test-project' . Str::uuid();
-        $response = $this->actingAs($this->users['normal'])->graphQL('
-            mutation CreateProject($input: CreateProjectInput!) {
-                createProject(input: $input) {
-                    id
-                    name
-                }
-            }
-        ', [
-            'input' => [
-                'name' => $name,
-                'description' => 'test',
-                'homeUrl' => 'https://cdash.org',
-                'visibility' => 'PUBLIC',
-                'authenticateSubmissions' => false,
-            ],
-        ]);
-
-        $project = Project::where('name', $name)->firstOrFail();
-
-        $response->assertExactJson([
-            'data' => [
-                'createProject' => [
-                    'id' => (string) $project->id,
-                    'name' => $name,
-                ],
-            ],
-        ]);
-
-        $project->delete();
-    }
-
-    public function testCreateProjectAdmin(): void
-    {
-        $name = 'test-project' . Str::uuid();
-        $response = $this->actingAs($this->users['admin'])->graphQL('
-            mutation CreateProject($input: CreateProjectInput!) {
-                createProject(input: $input) {
-                    id
-                    name
-                }
-            }
-        ', [
-            'input' => [
-                'name' => $name,
-                'description' => 'test',
-                'homeUrl' => 'https://cdash.org',
-                'visibility' => 'PUBLIC',
-                'authenticateSubmissions' => false,
-            ],
-        ]);
-
-        $project = Project::where('name', $name)->firstOrFail();
-
-        $response->assertExactJson([
-            'data' => [
-                'createProject' => [
-                    'id' => (string) $project->id,
-                    'name' => $name,
-                ],
-            ],
-        ]);
-
-        $project->delete();
     }
 
     public function testGetProjectAdministratorsNoUser(): void
@@ -899,153 +748,6 @@ class ProjectTypeTest extends TestCase
                 ],
             ],
         ]);
-    }
-
-    /**
-     * @return array{
-     *     array{
-     *         string,
-     *         string,
-     *         string,
-     *         bool
-     *     }
-     * }
-     */
-    public static function createProjectVisibilityRules(): array
-    {
-        return [
-            ['normal', 'PUBLIC', 'PUBLIC',  true],
-            ['normal', 'PROTECTED', 'PUBLIC', true],
-            ['normal', 'PRIVATE', 'PUBLIC', true],
-            ['normal', 'PUBLIC', 'PROTECTED', false],
-            ['normal', 'PROTECTED', 'PROTECTED', true],
-            ['normal', 'PRIVATE', 'PROTECTED', true],
-            ['normal', 'PUBLIC', 'PRIVATE', false],
-            ['normal', 'PROTECTED', 'PRIVATE', false],
-            ['normal', 'PRIVATE', 'PRIVATE', true],
-            ['admin', 'PUBLIC', 'PUBLIC', true],
-            ['admin', 'PROTECTED', 'PUBLIC', true],
-            ['admin', 'PRIVATE', 'PUBLIC', true],
-            ['admin', 'PUBLIC', 'PROTECTED', true],
-            ['admin', 'PROTECTED', 'PROTECTED', true],
-            ['admin', 'PRIVATE', 'PROTECTED', true],
-            ['admin', 'PUBLIC', 'PRIVATE', true],
-            ['admin', 'PROTECTED', 'PRIVATE',  true],
-            ['admin', 'PRIVATE', 'PRIVATE',  true],
-        ];
-    }
-
-    #[DataProvider('createProjectVisibilityRules')]
-    public function testCreateProjectMaxVisibility(string $user, string $visibility, string $max_visibility, bool $can_create): void
-    {
-        Config::set('cdash.user_create_projects', true);
-        Config::set('cdash.max_project_visibility', $max_visibility);
-
-        $name = 'test-project' . Str::uuid();
-        $response = $this->actingAs($this->users[$user])->graphQL('
-            mutation CreateProject($input: CreateProjectInput!) {
-                createProject(input: $input) {
-                    visibility
-                }
-            }
-        ', [
-            'input' => [
-                'name' => $name,
-                'description' => 'test',
-                'homeUrl' => 'https://cdash.org',
-                'visibility' => $visibility,
-                'authenticateSubmissions' => false,
-            ],
-        ]);
-
-        if ($can_create) {
-            $project = Project::where('name', $name)->firstOrFail();
-            $response->assertExactJson([
-                'data' => [
-                    'createProject' => [
-                        'visibility' => $visibility,
-                    ],
-                ],
-            ]);
-            $project->delete();
-        } else {
-            // A final check to ensure this project wasn't created anyway
-            $this->assertDatabaseMissing(Project::class, [
-                'name' => $name,
-            ]);
-            $response->assertGraphQLErrorMessage('Validation failed for the field [createProject].');
-        }
-    }
-
-    /**
-     * @return array{
-     *     array{
-     *         string,
-     *         bool,
-     *         bool,
-     *         bool
-     *     }
-     * }
-     */
-    public static function authenticatedSubmissionRules(): array
-    {
-        return [
-            ['normal', false, false,  true],
-            ['normal', true, false,  true],
-            ['normal', false, true,  false],
-            ['normal', true, true,  true],
-            // Instance admins can set any value
-            ['admin', false, false,  true],
-            ['admin', true, false,  true],
-            ['admin', false, true,  true],
-            ['admin', true, true,  true],
-        ];
-    }
-
-    #[DataProvider('authenticatedSubmissionRules')]
-    public function testRequireAuthenticatedSubmissions(
-        string $user,
-        bool $use_authenticated_submits,
-        bool $require_authenticated_submissions,
-        bool $result,
-    ): void {
-        Config::set('cdash.user_create_projects', true);
-        Config::set('cdash.require_authenticated_submissions', $require_authenticated_submissions);
-
-        $name = 'test-project' . Str::uuid();
-        $response = $this->actingAs($this->users[$user])->graphQL('
-            mutation CreateProject($input: CreateProjectInput!) {
-                createProject(input: $input) {
-                    authenticateSubmissions
-                }
-            }
-        ', [
-            'input' => [
-                'name' => $name,
-                'description' => 'test',
-                'homeUrl' => 'https://cdash.org',
-                'visibility' => 'PUBLIC',
-                'authenticateSubmissions' => $use_authenticated_submits,
-            ],
-        ]);
-
-        if ($result) {
-            $project = Project::where('name', $name)->firstOrFail();
-            $response->assertExactJson([
-                'data' => [
-                    'createProject' => [
-                        'authenticateSubmissions' => $use_authenticated_submits,
-                    ],
-                ],
-            ]);
-            $project->delete();
-        } else {
-            // A final check to ensure this project wasn't created anyway
-            $this->assertDatabaseMissing(Project::class, [
-                'name' => $name,
-            ]);
-            $response->assertGraphQLErrorMessage('Validation failed for the field [createProject].');
-        }
     }
 
     /**
